[Web] Fix : :fire: move MainButton.tsx to atoms directory and specify props type #99
